requirements.txt
tensorrt==8.6.1
librosa
tqdm
filetype
imageio
opencv_python_headless
scikit-image
cython
cuda-python
mediapipe
onnxruntime
imageio-ffmpeg
colored
polygraphy
numpy==2.0.1
flask-ngrok
pyngrok
gdown